another fibonacci trial program

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 int main(void)
 {
 	int count;
-	unsigned long long fib1 = 1, fib2 = 2, fib_next;
+	unsigned long fib1 = 1, fib2 = 2, fib_next;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		fib_next = fib1 + fib2;
 
-		printf(", %llu", fib_next);
+		printf(", %lu", fib_next);
 
 		fib1 = fib2;
 		fib2 = fib_next;
